+++ backend/src/chat/services/message.service.ts
@@ -13,27 +13,33 @@ export class MessageService {
   ) {}
 
   async create(message: IMessage): Promise<Message> {
-    const newMessage = this.messageRepository.create(message);
-    return this.messageRepository.save(newMessage);
+    if (message) {
+      const newMessage = this.messageRepository.create(message);
+      return this.messageRepository.save(newMessage);
+    }
   }
 
   async findMessageByChannel(channel: IChannel) {
-    return this.messageRepository
-      .createQueryBuilder('message')
-      .leftJoin('message.channel', 'channel')
-      .where('channel.id = :channelId', { channelId: channel.id })
-      .leftJoinAndSelect('message.user', 'user')
-      .leftJoinAndSelect('user.blockedUsers', 'message_blocked_users')
-      .orderBy('message.created_at', 'ASC')
-      .getMany();
+    if (channel) {
+      return this.messageRepository
+        .createQueryBuilder('message')
+        .leftJoin('message.channel', 'channel')
+        .where('channel.id = :channelId', { channelId: channel.id })
+        .leftJoinAndSelect('message.user', 'user')
+        .leftJoinAndSelect('user.blockedUsers', 'message_blocked_users')
+        .orderBy('message.created_at', 'ASC')
+        .getMany();
+    }
   }
 
   async deleteMessageByChannel(channel: IChannel) {
-    return this.messageRepository
-      .createQueryBuilder('message')
-      .leftJoin('message.channel', 'channel')
-      .where('channel.id = :channelId', { channelId: channel.id })
-      .delete()
-      .execute();
+    if (channel) {
+      return this.messageRepository
+        .createQueryBuilder('message')
+        .leftJoin('message.channel', 'channel')
+        .where('channel.id = :channelId', { channelId: channel.id })
+        .delete()
+        .execute();
+    }
   }
 }
